+++ CHANGELOG.md
@@ -1,5 +1,11 @@
 # CHANGELOG
 
+## Unreleased
+
+### Fixed
+
+* Fix throwing InvalidArgumentException on `Scraper::getApp()` when app page contains a `NULL` byte.
+
 ## 0.1.0 (2016-01-04)
 
 * Initial release.